requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
pandas==2.2.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
